% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_POO_thesaurus.R
\docType{class}
\name{Thesaurus}
\alias{Thesaurus}
\title{A Reference Class pour extraire le contenu du thesaurus des interactions de l'ANSM}
\value{
Permet de créer des instances de cette classe contenant diverses informations sur une mise à jour-thésaurus
}
\description{
Cette classe est utilisée pour transformer le contenu textuel (plain text obtenu par l'outil Tika à partir du PDF) du thésaurus
en contenu structuré (dataframe R).
}
\details{
Les fonctions servent à transformer le contenu du thesaurus d'un format textuel à un format structuré.
Elles n'ont pas d'utilité par la suite
}
\section{Fields}{

\describe{
\item{\code{thesaurus}}{: le contenu du thesaurus en texte brut}

\item{\code{lignes_entrees}}{: numéros des lignes où on trouvent une entréee}

\item{\code{regex_exclus}}{: si ces lignes commencent par ces regex, ce ne sont pas des entrées (ECG...)}

\item{\code{df}}{: dataframe contenant la liste des protagoniste A - B du thesaurus avec la description, le mécanisme et le niveau}

\item{\code{df_decompose}}{: pareil que df mais les familles sont remplacées par les molécules qu'elles contiennent}

\item{\code{manuellement}}{: subset de df où il existe plusieurs niveaux d'interaction selon le contexte.
L'idée est de traiter manuellement cette partie pour séparer les niveaux puis l'intégrer à df.}

\item{\code{mol_famille}}{: dataframe contenant les molécules et leurs familles.}

\item{\code{molecules_seules}}{: dataframe contenant la liste des molécules seules.}

\item{\code{mol}}{: vecteur contenant la liste des molécules du thésaurus}
}}

